requirements.txt
django-debug-toolbar==1.4
psycopg2==2.6.1
Django==1.9.5
